Change DetailsI to store client as raw pointer to prevent circular reference
This fixes a circular reference between client and details, which was causing the client to never be freed unless closeSession was called, which would free the Ice objects, and break the reference. The client controls destruction of details via closeSession, so there is no worry of a dangling pointer issue.